requirements.txt
amqp==5.0.9
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.6
django-admin-volt==0.0.3
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-embed-video==1.4.1
django-sendgrid-v5==1.2.0
djangorestframework==3.13.1
filelock==3.4.2
flower==1.0.0
fontawesome-free==5.15.4
future==0.18.2
gunicorn==20.1.0
honeybadger==0.8.0
html5lib==1.1
humanize==3.13.1
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
imgkit==1.2.2
kombu==5.2.2
lxml==4.8.0
moviepy==1.0.3
numpy==1.21.5
oauthlib==3.1.1
pandas==1.3.5
pdfkit==1.0.0
Pillow==8.4.0
platformdirs==2.4.1
plumbum==1.7.0
proglog==0.1.9
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.9.0
psycopg2-binary==2.9.2
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.3.0
PyPDF2==1.26.0
pypiwin32==223
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python-http-client==3.3.5
python-magic==0.4.24
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2021.1
pywin32==301
redis==4.0.2
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
sendgrid==6.9.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
toml==0.10.2
tornado==6.1
tqdm==4.62.3
Unipath==1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.13.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
winapps==0.2.0
wrapt==1.13.3
xhtml2pdf==0.2.5
